+++ forui/lib/src/widgets/select/multi/basic_select.dart
@@ -40,12 +40,15 @@ class _BasicSelect<T> extends FMultiSelect<T> {
     super.contentOffset,
     super.contentHideRegion,
     super.contentGroupId,
+    super.contentCutout,
+    super.contentCutoutBuilder,
     super.contentEmptyBuilder,
     super.contentScrollController,
     super.contentScrollHandles,
     super.contentPhysics,
     super.contentDivider,
     super.tagBuilder,
+    super.formFieldKey,
     super.key,
   }) : super._();