+++ forui/lib/src/widgets/select/multi/search_select.dart
@@ -48,12 +48,15 @@ class _SearchSelect<T> extends FMultiSelect<T> {
     super.contentOffset,
     super.contentHideRegion,
     super.contentGroupId,
+    super.contentCutout,
+    super.contentCutoutBuilder,
     super.contentEmptyBuilder,
     super.contentScrollController,
     super.contentScrollHandles,
     super.contentPhysics,
     super.contentDivider,
     super.tagBuilder,
+    super.formFieldKey,
     super.key,
   }) : super._();