+++ forui/lib/src/widgets/select/single/basic_select.dart
@@ -43,13 +43,15 @@ class _BasicSelect<T> extends FSelect<T> {
     super.contentOffset,
     super.contentHideRegion,
     super.contentGroupId,
+    super.contentCutout,
+    super.contentCutoutBuilder,
     super.autoHide,
     super.contentEmptyBuilder,
     super.contentScrollController,
     super.contentScrollHandles,
     super.contentPhysics,
     super.contentDivider,
-
+    super.formFieldKey,
     super.key,
   }) : super._();