+++ forui/lib/src/widgets/select/single/select.dart
@@ -202,6 +202,12 @@ abstract class FSelect<T> extends StatefulWidget with FFormFieldProperties<T> {
   /// {@macro forui.widgets.FPopover.groupId}
   final Object? contentGroupId;
 
+  /// {@macro forui.widgets.FPopover.cutout}
+  final bool contentCutout;
+
+  /// {@macro forui.widgets.FPopover.cutoutBuilder}
+  final void Function(Path path, Rect bounds) contentCutoutBuilder;
+
   /// True if the select should be automatically hidden after an item is selected. Defaults to false.
   final bool autoHide;
 
@@ -219,6 +225,9 @@ abstract class FSelect<T> extends StatefulWidget with FFormFieldProperties<T> {
 
   /// The divider used to separate the content items. Defaults to [FItemDivider.none].
   final FItemDivider contentDivider;
+
+  /// {@macro forui.foundation.doc_templates.formFieldKey}
+  final Key? formFieldKey;
 
   /// Creates a [FSelect] from the given [items].
   ///
@@ -266,12 +275,15 @@ abstract class FSelect<T> extends StatefulWidget with FFormFieldProperties<T> {
     Offset contentOffset = .zero,
     FPopoverHideRegion contentHideRegion = .excludeChild,
     Object? contentGroupId,
+    bool contentCutout = true,
+    void Function(Path path, Rect bounds) contentCutoutBuilder = FModalBarrier.defaultCutoutBuilder,
     bool autoHide = true,
     Widget Function(BuildContext context, FSelectStyle style) contentEmptyBuilder = defaultContentEmptyBuilder,
     ScrollController? contentScrollController,
     bool contentScrollHandles = false,
     ScrollPhysics contentPhysics = const ClampingScrollPhysics(),
     FItemDivider contentDivider = .none,
+    Key? formFieldKey,
     Key? key,
   }) {
     final inverse = {for (final MapEntry(:key, :value) in items.entries) value: key};
@@ -314,12 +326,15 @@ abstract class FSelect<T> extends StatefulWidget with FFormFieldProperties<T> {
       contentOffset: contentOffset,
       contentHideRegion: contentHideRegion,
       contentGroupId: contentGroupId,
+      contentCutout: contentCutout,
+      contentCutoutBuilder: contentCutoutBuilder,
       autoHide: autoHide,
       contentEmptyBuilder: contentEmptyBuilder,
       contentScrollController: contentScrollController,
       contentScrollHandles: contentScrollHandles,
       contentPhysics: contentPhysics,
       contentDivider: contentDivider,
+      formFieldKey: formFieldKey,
       key: key,
       children: [for (final MapEntry(:key, :value) in items.entries) .item(title: Text(key), value: value)],
     );
@@ -366,12 +381,15 @@ abstract class FSelect<T> extends StatefulWidget with FFormFieldProperties<T> {
     Offset contentOffset,
     FPopoverHideRegion contentHideRegion,
     Object? contentGroupId,
+    bool contentCutout,
+    void Function(Path path, Rect bounds) contentCutoutBuilder,
     bool autoHide,
     Widget Function(BuildContext context, FSelectStyle style) contentEmptyBuilder,
     ScrollController? contentScrollController,
     bool contentScrollHandles,
     ScrollPhysics contentPhysics,
     FItemDivider contentDivider,
+    Key? formFieldKey,
     Key? key,
   }) = _BasicSelect<T>;
 
@@ -434,12 +452,15 @@ abstract class FSelect<T> extends StatefulWidget with FFormFieldProperties<T> {
     Offset contentOffset = .zero,
     FPopoverHideRegion contentHideRegion = .excludeChild,
     Object? contentGroupId,
+    bool contentCutout = true,
+    void Function(Path path, Rect bounds) contentCutoutBuilder = FModalBarrier.defaultCutoutBuilder,
     bool autoHide = true,
     Widget Function(BuildContext context, FSelectStyle style) contentEmptyBuilder = defaultContentEmptyBuilder,
     ScrollController? contentScrollController,
     bool contentScrollHandles = false,
     ScrollPhysics contentPhysics = const ClampingScrollPhysics(),
     FItemDivider contentDivider = .none,
+    Key? formFieldKey,
     Key? key,
   }) {
     final inverse = {for (final MapEntry(:key, :value) in items.entries) value: key};
@@ -494,12 +515,15 @@ abstract class FSelect<T> extends StatefulWidget with FFormFieldProperties<T> {
       contentOffset: contentOffset,
       contentHideRegion: contentHideRegion,
       contentGroupId: contentGroupId,
+      contentCutout: contentCutout,
+      contentCutoutBuilder: contentCutoutBuilder,
       autoHide: autoHide,
       contentEmptyBuilder: contentEmptyBuilder,
       contentScrollController: contentScrollController,
       contentScrollHandles: contentScrollHandles,
       contentPhysics: contentPhysics,
       contentDivider: contentDivider,
+      formFieldKey: formFieldKey,
       key: key,
     );
   }
@@ -557,12 +581,15 @@ abstract class FSelect<T> extends StatefulWidget with FFormFieldProperties<T> {
     Offset contentOffset,
     FPopoverHideRegion contentHideRegion,
     Object? contentGroupId,
+    bool contentCutout,
+    void Function(Path path, Rect bounds) contentCutoutBuilder,
     bool autoHide,
     Widget Function(BuildContext context, FSelectStyle style) contentEmptyBuilder,
     ScrollController? contentScrollController,
     bool contentScrollHandles,
     ScrollPhysics contentPhysics,
     FItemDivider contentDivider,
+    Key? formFieldKey,
     Key? key,
   }) = _SearchSelect<T>;
 
@@ -605,12 +632,15 @@ abstract class FSelect<T> extends StatefulWidget with FFormFieldProperties<T> {
     this.contentOffset = .zero,
     this.contentHideRegion = .excludeChild,
     this.contentGroupId,
+    this.contentCutout = true,
+    this.contentCutoutBuilder = FModalBarrier.defaultCutoutBuilder,
     this.autoHide = true,
     this.contentEmptyBuilder = defaultContentEmptyBuilder,
     this.contentScrollController,
     this.contentScrollHandles = false,
     this.contentPhysics = const ClampingScrollPhysics(),
     this.contentDivider = .none,
+    this.formFieldKey,
     super.key,
   });
 
@@ -654,12 +684,15 @@ abstract class FSelect<T> extends StatefulWidget with FFormFieldProperties<T> {
       ..add(DiagnosticsProperty('contentOffset', contentOffset))
       ..add(EnumProperty('contentHideRegion', contentHideRegion))
       ..add(DiagnosticsProperty('contentGroupId', contentGroupId))
+      ..add(FlagProperty('contentCutout', value: contentCutout, ifTrue: 'cutout'))
+      ..add(ObjectFlagProperty.has('contentCutoutBuilder', contentCutoutBuilder))
       ..add(FlagProperty('autoHide', value: autoHide, ifTrue: 'autoHide'))
       ..add(ObjectFlagProperty.has('emptyBuilder', contentEmptyBuilder))
       ..add(DiagnosticsProperty('contentScrollController', contentScrollController))
       ..add(FlagProperty('contentScrollHandles', value: contentScrollHandles, ifTrue: 'contentScrollHandles'))
       ..add(DiagnosticsProperty('contentPhysics', contentPhysics))
-      ..add(EnumProperty('contentDivider', contentDivider));
+      ..add(EnumProperty('contentDivider', contentDivider))
+      ..add(DiagnosticsProperty('formFieldKey', formFieldKey));
   }
 }
 
@@ -752,6 +785,7 @@ abstract class _State<S extends FSelect<T>, T> extends State<S> with TickerProvi
     final localizations = FLocalizations.of(context) ?? FDefaultLocalizations();
 
     return Field<T>(
+      key: widget.formFieldKey,
       controller: _controller,
       enabled: widget.enabled,
       autovalidateMode: widget.autovalidateMode,
@@ -795,6 +829,8 @@ abstract class _State<S extends FSelect<T>, T> extends State<S> with TickerProvi
           offset: widget.contentOffset,
           hideRegion: widget.contentHideRegion,
           groupId: widget.contentGroupId,
+          cutout: widget.contentCutout,
+          cutoutBuilder: widget.contentCutoutBuilder,
           shortcuts: {const SingleActivator(.escape): _toggle},
           popoverBuilder: (_, popoverController) => TextFieldTapRegion(
             child: InheritedSelectController<T>(